// types/resume.ts
export interface Experience {
    company: string;
    role: string;
    period: string;
    desc: string[];
}

export interface Education {
    school: string;
    degree: string;
    gpa: string;
    period: string;
}

export interface PersonalInfo {
    name: string;
    role: string;
    subtitle: string;
    email: string;
    phone: string;
    location: string;
    summary: string;
}